Moved interactive_traversal
