+++ sympy/interactive/__init__.py
@@ -2,5 +2,7 @@
 
 from .printing import init_printing
 from .session import init_session
+from .traversal import interactive_traversal
 
-__all__ = ['init_printing', 'init_session']
+
+__all__ = ['init_printing', 'init_session', 'interactive_traversal']
